requirements.txt
selenium
chromedriver-binary-auto
pandas
beautifulsoup4
python-dotenv